made sure getch() return value is always stored in an int instead of a char to avoid loosing values over 127

+++ keypress.cpp
@@ -5,9 +5,9 @@
 keymap<action_id> keys;
 #endif
 
-long input()
+int input()
 {
- long ch = getch();
+ int ch = getch();
  switch (ch) {
   case KEY_UP:    return 'k';
   case KEY_LEFT:  return 'h';
